more protocol work

+++ framework/src/main/java/com/hellblazer/anubis/rst/udp/MsgTypes.java
@@ -1,8 +1,8 @@
 package com.hellblazer.anubis.rst.udp;
 
 public interface MsgTypes {
-    int ADD_CHILD = 0;
-    int REMOVE_CHILD = 1;
-    int SET_ROOT = 2;
-    int SET_COLOR = 3;
+    byte ADD_CHILD = 0;
+    byte REMOVE_CHILD = 1;
+    byte SET_ROOT = 2;
+    byte SET_COLOR = 3;
 }
